backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
torch==2.0.0
torchvision==0.15.0
Pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0 